# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Appointment:
    description: Appointment information
    properties:
      created_at:
        description: '@Description The time when the record was created'
        example: "2023-01-01T00:00:00Z"
        type: string
      date_time:
        description: '@Description The date and time of the appointment'
        example: "2023-01-01T10:00:00Z"
        type: string
      deleted_at:
        description: '@Description The time when the record was deleted (soft delete)'
        example: "2023-01-01T00:00:00Z"
        type: string
      doctor_id:
        description: '@Description The ID of the doctor'
        example: 2
        type: integer
      id:
        description: '@Description The unique identifier for the record'
        example: 1
        type: integer
      patient_id:
        description: '@Description The ID of the patient'
        example: 1
        type: integer
      reason:
        description: '@Description The reason for the appointment'
        example: Τακτικός έλεγχος
        type: string
      status:
        description: '@Description The status of the appointment'
        example: Created
        type: string
      updated_at:
        description: '@Description The time when the record was last updated'
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
...
  models.LoginCredentials:
    description: Login credentials
    properties:
      email:
        description: '@Description The user''s email address'
        example: georgios.papadopoulos@example.com
        type: string
      password:
        description: '@Description The user''s password'
        example: strongPassword123!
        type: string
    required:
    - email
    - password
    type: object
  models.LoginResponse:
    description: Login response
    properties:
      token:
        description: '@Description JWT token for authentication'
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  models.User:
    description: User information
    properties:
      created_at:
        description: '@Description The time when the record was created'
        example: "2023-01-01T00:00:00Z"
        type: string
      deleted_at:
        description: '@Description The time when the record was deleted (soft delete)'
        example: "2023-01-01T00:00:00Z"
        type: string
      email:
        description: '@Description The user''s email address'
        example: georgios.papadopoulos@example.com
        type: string
      first_name:
        description: '@Description The user''s first name'
        example: Γεώργιος
        type: string
      id:
        description: '@Description The unique identifier for the record'
        example: 1
        type: integer
      last_name:
        description: '@Description The user''s last name'
        example: Παπαδόπουλος
        type: string
      role:
        description: '@Description The user''s role in the system'
        example: patient
        type: string
      updated_at:
        description: '@Description The time when the record was last updated'
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  models.UserRegistration:
    description: User registration information
    properties:
      email:
        description: '@Description The user''s email address'
        example: georgios.papadopoulos@example.com
        type: string
      first_name:
        description: '@Description The user''s first name'
        example: Γεώργιος
        type: string
      last_name:
        description: '@Description The user''s last name'
        example: Παπαδόπουλος
        type: string
      password:
        description: '@Description The user''s password'
        example: strongPassword123!
        minLength: 8
        type: string
      role:
        description: '@Description The user''s role in the system'
        enum:
        - patient
        - doctor
        - receptionist
        example: patient
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - role
    type: object
  utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a clinic management server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Clinic Management API
  version: "1.0"
paths:
  /appointments:
    get:
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: startDate
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: endDate
        type: string
      - description: Patient ID
        in: query
        name: patientID
        type: integer
      - description: Doctor ID
        in: query
        name: doctorID
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Appointment'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: List appointments
      tags:
      - Appointments
    post:
      parameters:
      - description: Appointment data
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/models.Appointment'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Appointment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Create appointment
      tags:
      - Appointments
  /appointments/{id}:
    delete:
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete appointment
      tags:
      - Appointments
    get:
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Appointment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get appointment
      tags:
      - Appointments
    put:
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated appointment data
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/models.Appointment'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Appointment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update appointment
      tags:
      - Appointments
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: Login Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.LoginCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: User login
      tags:
      - Authentication
  /auth/logout:
    post:
      description: Logout the current user (invalidate the token)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with the provided details
      parameters:
      - description: User Registration Info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserRegistration'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Register a new user
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"